+++ src/Routes.js
@@ -9,7 +9,6 @@ import unreviewed from "./connectors/unreviewedProposals";
 import admin from "./connectors/admin";
 import newProposal from "./connectors/newProposal";
 import editProposal from "./connectors/editProposal";
-import singleComment from "./connectors/singleComment";
 
 import Logout from "./components/LogoutPage";
 import UserLookup from "./components/UserLookupPage";
@@ -58,7 +57,7 @@ class Routes extends Component {
         <AdminAuthenticatedRoute path="/admin/users" component={UserLookup} />
         <Route path="/user/:userId" component={userDetail(UserDetail)} />
         <Route path="/proposals/:token" component={proposalDetail(ProposalDetail)} exact />
-        <Route path="/proposals/:token/comments/:commentid" component={singleComment(ProposalDetail)} />
+        <Route path="/proposals/:token/comments/:commentid" component={proposalDetail(ProposalDetail)} />
         <Route path="/500" component={ErrorPage} />
         <Route path="*" component={NotFound} />
       </Switch>
