+++ src/components/snew/CommentArea.js
@@ -18,36 +18,32 @@ const CommentArea = ({
   proposal,
   onSetCommentsSortOption,
   commentsSortOption,
+  commentid,
+  onViewAllClick,
+  token,
   ...props
-}) => (
-  Object.keys(proposal).length === 0 ||
+}) => (Object.keys(proposal).length === 0 ||
   proposal.status === PROPOSAL_STATUS_UNREVIEWED ||
   proposal.status === PROPOSAL_STATUS_UNREVIEWED_CHANGES ||
-  proposal.status === PROPOSAL_STATUS_CENSORED ? null :
+  proposal.status === PROPOSAL_STATUS_CENSORED ? null : (
     <CommentAreaBase {...{
       ...props,
+      singleThread: commentid >= 0 ? (
+        <span>
+          Single comment thread.{" "}
+          <a href={`proposals/${token}`} onClick={onViewAllClick}>View all</a>
+        </span>
+      ) : null,
       locked: !loggedInAsEmail,
       name: TOP_LEVEL_COMMENT_PARENTID,
       num_comments: comments.length,
-      SorterComponent: () => comments && comments.length > 0 ? (
-        <div className="comments-sort" >
-          <span className="">Sort by:</span>
-          <Select
-            onKeyDown={(e) => e.keyCode === 8 && e.preventDefault()}
-            classNamePrefix="sort-select"
-            isSearchable={false}
-            isClearable={false}
-            escapeClearsValue={false}
-            value={commentsSortOption}
-            onChange={onSetCommentsSortOption}
-            options={[ SORT_BY_NEW, SORT_BY_OLD, SORT_BY_TOP ]
-              .map(op => ({ value: op, label: op }))
-            }
-          />
-        </div>
-      ) : null,
+      SorterComponent: () => comments && comments.length > 0 ? (<div className="comments-sort">
+        <span className="">Sort by:</span>
+        <Select onKeyDown={(e) => e.keyCode === 8 && e.preventDefault()} classNamePrefix="sort-select" isSearchable={false} isClearable={false} escapeClearsValue={false} value={commentsSortOption} onChange={onSetCommentsSortOption} options={[ SORT_BY_NEW, SORT_BY_OLD, SORT_BY_TOP ]
+          .map(op => ({ value: op, label: op }))} />
+      </div>) : null,
       hideSortOptions: false
     }} />
-);
+  ));
 
 export default proposalConnector(CommentArea);
